move templates, error handle init
