+++ Cargo.toml
@@ -16,3 +16,4 @@ headers = "0.3.9"
 serde_with = "3.3.0"
 futures-util = "0.3.28"
 axum-flash = "0.7.0"
+futures-core = "0.3.28"
